fix: Migrate `sql_json_merge()` related test cases to corresponding package
